refactor: dto 생성 시점을 domain과 분리하도록 변경
